Add <space> dot Puppet => Alignment
I finally figured out a good way to align => (fat commas) when working
in Puppet, this commit adds a function called SelectIndent to select the
current indent level, and then a visual block map to execute Align =>
when the period key is pressed.

I need to figure out how to save the current cursor position when
executing the SelectIndent function, but this shouldn't be a huge
problem to overcome.