Keep updating Conque terminals while unfocused.
Without this patch the terminal window running watchr inside of conque
doesn't automatically update.  This is a problem because the whole point
of conque for me is to automatically see the results of the spec tests
as I code, live.

This patch fixes the problem by configuring conque to automatically
update the buffer even when the window is unfocused.  This setting is:

    let g:ConqueTerm_ReadUnfocused = 1