Run RSpec2 tests when writing a *_spec.rb file
Without this patch it's still a bit cumbersome to run the test suite.
This patch addresses the problem by automatically running `make %` when
a buffer named *_spec.rb is saved.  This expects the compiler to be set
to rspec for ruby files.  If the behavior examples fail, the cursor will
automatically be repositioned to the failure.