Fix :RelatedSpecOpen and RelatedSpec() function
Without this patch the :RelatedSpecOpen and RelatedSpec function don't
work with the convention we commonly follow at Puppet Labs of putting
spec files in spec/unit/foo_spec.rb for a file in lib/foo.rb.  The root
cause of the problem is the assumption the RelatedSpec() function makes
about where spec files are located.

This patch fixes the problem by trimming both lib/ and app/ and the edge
case of 'lib' and 'app'.  The patch also adds the additional search
directory of `spec/unit` to the list of places to look.

Usage: `:RelatedSpecOpen` while in `lib/puppet.rb` opens
`spec/unit/puppet_spec.rb` if it exists.