Fix the spec finder
Without this patch CTRL-R doesn't work to open the related spec in Vim.
This is annoying.  The root cause is that expand("%:h") does not include
the trailing slash, but we match against the pattern ^lib/puppet/$.

This patch fixes the problem by removing the trailing /.  It also makes
things work with facter.