Make the RelatedSpec() project aware
Without this patch the related spec function doesn't work well when the
current working directory is not in a specific project, but rather
outside in the parent directory of a bunch of projects.

This patch addresses the problem by locating the project directory for
the given file and basing the related spec search off the project
directory.